backend/requirements.txt
langchain
langchain_community
azure-cognitiveservices-speech
langchain_core
langchain_openai
python-dotenv
flask
flask_cors
tiktoken
qdrant-client
openai
gunicorn